Date range comparisons
- Split out the date time popover as a separate component so it can be controlled by two different date range buttons (second date range is currently not used by anything).
- Add local state that is only "applied" on close to reduce thrashing.
- Add removable second button.
